Changed TYPE_PREFIXES to DATA_WRAPPERS and added settings, group, subgroup, item and ignore to the keys and changed the values to be a list of 2 values; one for the start of the string and one for the end.
Made a new function _check_string_for_wrappers to make code more readable and less repetitve by doing startswith and endswith checks in it.

Made the load function read types and also used recursive functions to be able to read nested loops indefinitely; well until the recursive error comes up.

Added update and new functions with the data set functions for changing the stored data.

Made a testing file that tests all the features instead of random nelson and stuff

Updated the filesavesystem class to have a __str__ so that it prints all variables when printed.

Sorry for huge commit